Created shared view model (crashing)
